server/requirements.txt
fastapi
uvicorn
pymongo
apscheduler
httpx
requests
beautifulsoup4
pandas
openpyxl
python-jose[cryptography]
passlib[argon2]
python-multipart
selenium
webdriver-manager
python-dotenv
